Migrate instance id APIs to EntityId for Unity 6.5

+++ src/UniCli.Unity/Packages/com.yucchiy.unicli-server/Editor/Handlers/Animator/SetParameterHandler.cs
@@ -118,7 +118,7 @@ namespace UniCli.Server.Editor.Handlers
     [Serializable]
     public class AnimatorSetParameterRequest
     {
-        public int instanceId;
+        public long instanceId;
         public string path = "";
         public string name = "";
         public string value = "";
